% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{annofile}
\alias{annofile}
\title{Protein annotation file.}
\format{
An object of class \code{character} of length 278605.
}
\usage{
annofile
}
\description{
A dataset containing the protein annotation file.THe annotation file could be read by readLines.
The dataset can be generated by querying protein sequences against protein domain database by rpsblast and rpsbproc.
}
\keyword{datasets}
