% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreDat.r
\name{CreDat}
\alias{CreDat}
\title{a function to merge gff file and annotation file}
\usage{
CreDat(gff, annofile)
}
\value{
a dataframe
}
\description{
a function to merge gff file and protein domain annotation file
}
\examples{
data(gff)
data(annofile)
mergedgff <- CreDat(gff,annofile)
}
\author{
Hongwei Wang <\email{whweve@163.com}>
}
