% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runVPDapp.r
\name{runVPDapp}
\alias{runVPDapp}
\title{a function to initiate shiny app VisDom}
\usage{
runVPDapp()
}
\value{
a shiny app
}
\description{
to  initiate a shiny app designed to plot gene structure and protein domain for given gene across species
}
\author{
Hongwei Wang <\email{whweve@163.com}>
}
